#  testing: Add the p99 quantile of write latency to `pageworkload`