--- a/dbms/src/Common/TiFlashMetrics.h
+++ b/dbms/src/Common/TiFlashMetrics.h
@@ -1061,9 +1061,9 @@ private:
     std::unordered_map<String, std::vector<T *>> resource_group_metrics_map;
 };
 
-namespace PS::tests
+namespace tests
 {
-class StressWorkload;
+struct TiFlashMetricsHelper;
 }
 
 /// Centralized registry of TiFlash metrics.
@@ -1119,7 +1119,7 @@ public:
     DISALLOW_COPY_AND_MOVE(TiFlashMetrics);
 
     friend class MetricsPrometheus;
-    friend class DB::PS::tests::StressWorkload;
+    friend struct DB::tests::TiFlashMetricsHelper;
 };
 
 #define MAKE_METRIC_ENUM_M(family_name, help, type, ...) \