--- a/dbms/src/Storages/Page/workload/TiFlashMetricsHelper.h
+++ b/dbms/src/Storages/Page/workload/TiFlashMetricsHelper.h
@@ -1,0 +1,45 @@
+// Copyright 2023 PingCAP, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include <Common/FmtUtils.h>
+#include <Common/TiFlashMetrics.h>
+#include <common/types.h>
+#include <prometheus/metric_type.h>
+
+#include <magic_enum.hpp>
+#include <unordered_set>
+
+namespace DB::tests
+{
+struct TiFlashMetricsHelper
+{
+public:
+    // "name-label" -> histogram
+    static std::unordered_map<String, prometheus::ClientMetric::Histogram> //
+    collectHistorgrams(const std::unordered_set<String> & names);
+
+    struct HistStats
+    {
+        double p99ms = 0.0;
+        double p999ms = 0.0;
+        double avgms = 0.0;
+    };
+
+    static HistStats histogramStats(const prometheus::ClientMetric::Histogram & hist);
+
+    static double histogramQuantile(const prometheus::ClientMetric::Histogram & hist, double q);
+    static double histogramAvg(const prometheus::ClientMetric::Histogram & hist);
+};
+
+} // namespace DB::tests