### What problem does this PR solve?

Issue Number: ref https://github.com/pingcap/tiflash/issues/8650

Problem Summary:

### What is changed and how it works?

Following testing after https://github.com/pingcap/tiflash/pull/8651. Run `./tiflash pageworkload --color false` for 3 times

-- | read_speed (GB/s) | write_speed(GB/s)
-- | -- | --
without optimization | 28.83633333 | 0.9126666667
with optimization | 33.02466667 | 0.9873333333


--  | blob_write-p999 | blob_write-p99 | blob_write-avg | total-p999 | total-avg
-- | -- | -- | -- | -- | --
without optimization | 99.78 | 45.47233333 | 8.125333333 | 118.4893333 | 9.66
with optimization | 28.968 | 16.99433333 | 7.356 | 37.537 | 8.836666667

We can see that the read_speed increase about 14%, write_speed increase about 8%. And p999 write latency down from 118.49ms to 37.53ms, avg write latency down from 9.66ms to 8.83ms


Details logging [bench.tar.gz](https://github.com/pingcap/tiflash/files/13821023/bench.tar.gz)

stdout example after this PR
```
Workload summary: {"memory":{"avg":"31002413.650","latest":165804970,"top":444598083},"read":{"bytes_read":19552,"pages":10093325,"read_speed":"32.587"},"snapshots":{"avg":"550135.108","latest":907229,"top":924210},"write":{"bytes_written":596,"pages":305542,"write_speed":"0.995"},"write_latency":{"blob_write":{"avg":"7.343","p99":"17.031","p999":"27.791"},"choose_stat":{"avg":"0.050","p99":"0.077","p999":"0.134"},"commit":{"avg":"0.050","p99":"0.050","p999":"0.050"},"latch":{"avg":"0.050","p99":"0.050","p999":"0.050"},"search_pos":{"avg":"0.050","p99":"0.050","p999":"0.050"},"total":{"avg":"8.770","p99":"22.343","p999":"35.143"},"wait_in_group":{"avg":"0.050","p99":"0.283","p999":"1.661"},"wal":{"avg":"0.203","p99":"1.537","p999":"3.107"}}}

```

### Check List

Tests <!-- At least one of them must be included. -->

- [ ] Unit test
- [ ] Integration test
- [x] Manual test (add detailed scripts or steps below)
- [ ] No code

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

Documentation

- [ ] Affects user behaviors
- [ ] Contains syntax changes
- [ ] Contains variable changes
- [ ] Contains experimental features
- [ ] Changes MySQL compatibility

### Release note

<!-- bugfix or new feature needs a release note -->

```release-note
None
```